feat(ui-tars): refact remote logic
