+++ apps/ui-tars/src/main/services/runAgent.ts
@@ -10,11 +10,11 @@ import { type ConversationWithSoM } from '@main/shared/types';
 import { GUIAgent, type GUIAgentConfig } from '@ui-tars/sdk';
 import { markClickPosition } from '@main/utils/image';
 import { UTIOService } from '@main/services/utio';
+import { NutJSElectronOperator } from '../agent/operator';
 import {
   createRemoteBrowserOperator,
-  NutJSElectronOperator,
   RemoteComputerOperator,
-} from '../agent/operator';
+} from '../remote/operators';
 import {
   DefaultBrowserOperator,
   RemoteBrowserOperator,
